refactored get access conditions
