+++ sdk/storage/azure-storage-blob/azure/storage/blob/_serialize.py
@@ -26,7 +26,7 @@ from ._generated.models import (
     JsonTextConfiguration,
     QueryFormatType,
     BlobTag,
-    BlobTags
+    BlobTags, LeaseAccessConditions
 )
 
 
@@ -177,3 +177,12 @@ def serialize_query_format(formater):
     else:
         raise TypeError("Format must be DelimitedTextDialect or DelimitedJsonDialect.")
     return QuerySerialization(format=qq_format)
+
+
+def get_access_conditions(lease):
+    # type: (Optional[Union[BlobLeaseClient, str]]) -> Union[LeaseAccessConditions, None]
+    try:
+        lease_id = lease.id # type: ignore
+    except AttributeError:
+        lease_id = lease # type: ignore
+    return LeaseAccessConditions(lease_id=lease_id) if lease_id else None
