+++ sdk/storage/azure-storage-blob/azure/storage/blob/aio/_container_client_async.py
@@ -30,9 +30,8 @@ from .._generated.models import (
     StorageErrorException,
     SignedIdentifier)
 from .._deserialize import deserialize_container_properties
-from .._serialize import get_modify_conditions, get_container_cpk_scope_info, get_api_version
+from .._serialize import get_modify_conditions, get_container_cpk_scope_info, get_api_version, get_access_conditions
 from .._container_client import ContainerClient as ContainerClientBase, _get_blob_name
-from .._lease import get_access_conditions
 from .._models import ContainerProperties, BlobType, BlobProperties  # pylint: disable=unused-import
 from ._list_blobs_helper import BlobPropertiesPaged, BlobPrefix
 from ._lease_async import BlobLeaseClient
